pwm: lpss: fix base_unit calculation for PWM frequency
The base_unit calculation applies an offset of 0x2 which adds significant
error for lower frequencies and doesn't appear to be justified - rounding
the division result gives a correct value.

Also, the upper limit check for base_unit is off-by-one; the upper nibble of
base_unit is invalid if >=128 according to the Table 88 in the Z8000 Processor
Series Datasheet Volume 1 (Rev. 2).

Signed-off-by: Dan O'Donovan <dan@emutex.com>